HeurBot v23 * 2P spawn logic: have 10 ships more than the opponent * Mark dropoffs neighbourhood (radius 2) as safe, always

+++ src/random_search.cpp
@@ -111,7 +111,7 @@ Solution Search::get_rnd()
 {
     Solution sol = Solution();
 
-    for (size_t i = 0; i < engine->game->me->ships.size(); i++)
+    for (int i = 0; i < engine->game->me->ships.size(); i++)
     {
         ShipSolution ship_sol = ShipSolution();
